Refactor: web3 provider 설정 분리
