+++ client/src/index.js
@@ -5,31 +5,13 @@ import reportWebVitals from "./reportWebVitals";
 import App from "./App";
 import store from "./store/store";
 
-import {
-  ThirdwebProvider,
-  metamaskWallet,
-  coinbaseWallet,
-  walletConnectV1,
-} from "@thirdweb-dev/react";
-import { Sepolia } from "@thirdweb-dev/chains";
-
 import "./index.css";
+
+import { ThirdwebProvider } from "@thirdweb-dev/react";
 import { SnackbarProvider } from "notistack";
 import { ThemeProvider, createTheme } from "@mui/material";
+import { Sepolia, supportedWallets, sdkOptions } from "./config/dappOptions";
 import customTheme from "./config/customTheme";
-
-const sdkOptions = {
-  gasSettings: {
-    speed: "fastest",
-  },
-  // gasless: {},
-};
-// const smartWalletConfig = {
-//   factoryAddress: "",
-//   gasless: true,
-//   thirdwebApiKey: process.env.REACT_APP_THIRDWEB_API_KEY,
-//   personalWallets: [metamaskWallet(), coinbaseWallet(), walletConnectV1()],
-// };
 
 const ThemedApp = () => {
   const basicTheme = createTheme({});
@@ -47,7 +29,7 @@ root.render(
   <Provider store={store}>
     <ThirdwebProvider
       activeChain={Sepolia}
-      supportedWallets={[metamaskWallet(), coinbaseWallet(), walletConnectV1()]}
+      supportedWallets={supportedWallets}
       sdkOptions={sdkOptions}
     >
       <SnackbarProvider
